server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdf2image==1.17.0
Pillow==10.4.0
aiofiles==23.2.1
typing-extensions>=4.8.0
openai-whisper==20231117
google-genai>=0.5.0
python-dotenv>=1.0.1
